dom/base/test/meta_viewport/test_meta_viewport_auto_size_by_device_height.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>device-height enables autoSize</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="height=device-height">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>height=device-height</p>
  <script type="application/javascript">
    "use strict";

    add_task(async function device_height() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      ok(info.autoSize, "device-height should enable autoSize");
    });
  </script>
</body>
</html>
